setting: 테스트 환경 설정
- jest, babel 설치
- package.json 스크립트 작성
- tsconfig.json의 root 수정
